[RTD-193] rename controller to be self-explanatory
